+++ src/lib.rs
@@ -2,7 +2,6 @@ mod error;
 
 pub mod capturer;
 pub mod duplication_context;
-pub mod manager;
 pub mod model;
 pub mod utils;